+++ common/src/main/kotlin/net/spaceeye/vmod/gui/ScreenWindow.kt
@@ -10,10 +10,11 @@ import net.minecraft.client.gui.GuiGraphics
 import net.spaceeye.vmod.gui.additions.ErrorAddition
 import net.spaceeye.vmod.gui.additions.HUDAddition
 import net.spaceeye.vmod.gui.additions.InfoAddition
+import net.spaceeye.vmod.gui.additions.PresetsAddition
 import net.spaceeye.vmod.gui.additions.VEntityChangerWorldMenu
 
 
-//TODO i don't like it
+//TODO maybe make it a class so that i can hot reload when elementa crashes
 object ScreenWindow: WindowScreen(ElementaVersion.V8, drawDefaultBackground = false) {
     private val linearExtensions = mutableListOf<ScreenWindowAddition>()
     private val _extensions = mutableSetOf<ScreenWindowAddition>()
@@ -67,5 +68,6 @@ object ScreenWindow: WindowScreen(ElementaVersion.V8, drawDefaultBackground = fa
         addScreenAddition { ErrorAddition() }
         addScreenAddition { VEntityChangerWorldMenu }
         addScreenAddition { InfoAddition }
+        addScreenAddition { PresetsAddition }
     }
 }