+++ common/src/main/kotlin/net/spaceeye/vmod/toolgun/ToolgunItem.kt
@@ -14,6 +14,7 @@ import net.spaceeye.vmod.events.PersistentEvents
 import net.spaceeye.vmod.toolgun.ClientToolGunState.playerIsUsingToolgun
 import org.lwjgl.glfw.GLFW
 
+//TODO why am i defining events here, move them to ClientToolgunState
 class ToolgunItem: Item(Properties().stacksTo(1).`arch$tab`(VMItems.TAB)) {
     companion object {
         @JvmStatic