+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.6.0
+* Added delete button to VEntityChanger
+* Added Settings Presets
+* Fixed rare crash when raycasting
+* Rolled back schematic renderer change cuz it didn't work
+* Fixed phys bearing schem compat not working
+
 # 1.5.1
 * Fixed incompat with control craft
 * Changed schematic renderer